requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
line-bot-sdk==3.14.3
google-generativeai==0.8.4
sentence-transformers==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
gunicorn==21.2.0
